[63] 基本的な階層構造を修正
