+++ app/src/main/java/com/example/tokumemo/ui/home/HomeMenuRecyclerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.tokumemo
+package com.example.tokumemo.ui.home
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.tokumemo.R
 
 class HomeMenuRecyclerAdapter(private val fields: List<HomeListData>): RecyclerView.Adapter<HomeMenuRecyclerAdapter.ViewHolder>() {
