+++ app/src/main/java/com/example/tokumemo/utility/Extension.kt
@@ -1,4 +1,4 @@
-package com.example.tokumemo
+package com.example.tokumemo.utility
 
 inline fun <T> guard(value: T?, ifNull: () -> Unit): T {
     if (value != null) return value
