+++ app/src/main/java/com/example/tokumemo/utility/FirstDialogFragment.kt
@@ -1,3 +1,5 @@
+package com.example.tokumemo.utility
+
 import android.app.AlertDialog
 import android.app.Dialog
 import android.content.DialogInterface
