+++ app/src/main/res/layout/fragment_news.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center"
-    tools:context=".NewsFragment">
+    tools:context=".ui.news.NewsFragment">
 
     <TextView
         android:id="@+id/club_lists_title_bar"
