+++ app/src/main/res/layout/fragment_settings.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_gravity="center"
     android:background="@color/back_ground_gray_color"
-    tools:context=".SettingsFragment">
+    tools:context=".ui.settings.SettingsFragment">
 
     <TextView
         android:id="@+id/settings_title"
